added example page

+++ xmpp-client/src/main/java/de/tu_berlin/cit/intercloud/xmpp/client/occi/representation/Attribute.java
@@ -144,7 +144,7 @@ public class Attribute implements Serializable {
      */
 
     public boolean isFloat() {
-        return Type.INTEGER.equals(this.type);
+        return Type.FLOAT.equals(this.type);
     }
 
     public void setFloat(Float f) {
